third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/container-type-parsing.html

<!DOCTYPE html>
<title>Parsing of container-type</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<div id="target"></div>
<script>
test_valid_value('container-type', 'initial');
test_valid_value('container-type', 'inherit');
test_valid_value('container-type', 'unset');
test_valid_value('container-type', 'revert');
test_valid_value('container-type', 'none');
test_valid_value('container-type', 'inline-size');
test_valid_value('container-type', 'block-size');
test_valid_value('container-type', 'size');

test_invalid_value('container-type', 'none none');
test_invalid_value('container-type', 'none inline-size');
test_invalid_value('container-type', 'inline-size none');
test_invalid_value('container-type', 'inline-size inline-size');
test_invalid_value('container-type', 'inline-size block-size');
test_invalid_value('container-type', 'block-size inline-size');
test_invalid_value('container-type', 'size inline-size');
test_invalid_value('container-type', 'inline-size size');
test_invalid_value('container-type', 'none, none');
test_invalid_value('container-type', 'foo');
test_invalid_value('container-type', '"foo"');
test_invalid_value('container-type', 'foo, bar');
test_invalid_value('container-type', '#fff');
test_invalid_value('container-type', '1px');
test_invalid_value('container-type', 'default');
test_invalid_value('container-type', 'size nonsense');
</script>
